% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRFMScores.R
\name{RFMfunction}
\alias{RFMfunction}
\title{RFMfunction}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{data}{- some text}

\item{r}{- weight of recency}
}
\value{
Returns a data.table containing stuff
}
\description{
Some Description text
}
\details{
\code{data} some more infos ...
}
